pass models and serializers of vacancy from apps.user to apps.application
